EA: deriner commit avant de commencer le livrable deux'

+++ Chess/modeleDame.h
@@ -7,7 +7,7 @@ namespace modele {
 class Dame : public Piece
 {
 public:
-    explicit Dame(QObject *parent = nullptr, pair<uint8_t,uint8_t> position = {0,0});
+    explicit Dame(QObject *parent = nullptr, pair<int,int> position = {0,0});
 public slots:
     void mettreAJourPositionsValides() override;
 };
